Update README: run `mdf checkout -f` on new setup
When setting up on a new machine, you may need to force the checkout.
This will overwrite local files with the same names as those tracked in
.mdf, which should be desried behavior for a new machine anyway. I
noticed this was needed when I tested on a local server and the files
didn't initially show up in my $HOME when I ran `mdf checkout`.