Update .zshrc: make PROMPT_TIMING check consistent
The first condition at the start of the file used the `-v` operater to
determine if the PROMPT_TIMING env var is set. The final condition used
`-n` to determine if the value of the env var was non-empty. I changed
this to `-v` so that the logic is consistent with the first check.