Update zsh_key_binds: change condition for macOS
The goal was to change my keybind for Option (or Alt) + the arrow keys
to send a jump forward by word and jump backward by word signal if I
were on macOS. But really, it's the terminal emulator I'm connected with
that matters. I could be connected to a linux server with iTerm2 as my
ssh client. I still want to be able to use those keys to jump by word.
So in this case, I check if the LC_TERMINAL variable matches iTerm2 and
set the binding.